Move eth_getProof code into evm crate
