+++ crates/evm/src/rpc_helpers/mod.rs
@@ -1,24 +1,27 @@
 use std::collections::HashMap;
 
-use alloy_primitives::{keccak256, Address};
+use alloy_consensus::constants::KECCAK_EMPTY;
+use alloy_primitives::{keccak256, Address, Bytes};
 use alloy_rpc_types::state::AccountOverride;
-use alloy_rpc_types::BlockOverrides;
+use alloy_rpc_types::{BlockOverrides, EIP1186AccountProofResponse, EIP1186StorageProof};
+use alloy_serde::JsonStorageKey;
 pub use filter::*;
 pub use log_utils::*;
 use reth_rpc_eth_types::{EthApiError, EthResult};
+use revm::context::BlockEnv;
 use revm::Database;
 
 mod filter;
 mod log_utils;
 mod tracing_utils;
 
-#[cfg(feature = "native")]
-use revm::context::BlockEnv;
+use sov_modules_api::{StateMapAccessor, WorkingSet};
+use sov_state::storage::{NativeStorage, StateCodec, StorageKey};
 pub(crate) use tracing_utils::*;
 
 use crate::db::EvmDb;
+use crate::Evm;
 
-#[cfg(feature = "native")]
 /// Applies all instances [`AccountOverride`] to the [`EvmDb`].
 pub(crate) fn apply_state_overrides<C: sov_modules_api::Context>(
     state_overrides: HashMap<Address, AccountOverride, alloy_primitives::map::FbBuildHasher<20>>,
@@ -31,7 +34,6 @@ pub(crate) fn apply_state_overrides<C: sov_modules_api::Context>(
     Ok(())
 }
 
-#[cfg(feature = "native")]
 /// Applies a single [`AccountOverride`] to the [`EvmDb`].
 pub(crate) fn apply_account_override<C: sov_modules_api::Context>(
     account: Address,
@@ -73,7 +75,6 @@ pub(crate) fn apply_account_override<C: sov_modules_api::Context>(
     Ok(())
 }
 
-#[cfg(feature = "native")]
 /// Applies all instances of [`BlockOverride`] to the [`EvmDb`].
 pub(crate) fn apply_block_overrides<C: sov_modules_api::Context>(
     block_env: &mut BlockEnv,
@@ -114,5 +115,130 @@ pub(crate) fn apply_block_overrides<C: sov_modules_api::Context>(
     }
     if let Some(base_fee) = base_fee {
         block_env.basefee = base_fee.saturating_to();
+    }
+}
+
+/// Returns zkproof by EIP-1186 (eth_getProof).
+// `Evm`` and `working_set`` must be rewind to the specified `version``.
+pub fn generate_eth_proof<C: sov_modules_api::Context>(
+    evm: &Evm<C>,
+    address: Address,
+    keys: Vec<JsonStorageKey>,
+    version: u64,
+    working_set: &mut WorkingSet<C::Storage>,
+) -> EIP1186AccountProofResponse
+where
+    C::Storage: NativeStorage,
+{
+    let root_hash = working_set
+        .get_root_hash(version)
+        .expect("Root hash must exist for all blocks");
+
+    let account = evm.account_info(&address, working_set).unwrap_or_default();
+    let balance = account.balance;
+    let nonce = account.nonce;
+    let code_hash = account.code_hash.unwrap_or(KECCAK_EMPTY);
+
+    fn generate_account_proof<C>(
+        evm: &Evm<C>,
+        account: &Address,
+        version: u64,
+        working_set: &mut WorkingSet<C::Storage>,
+    ) -> Vec<Bytes>
+    where
+        C: sov_modules_api::Context,
+        C::Storage: NativeStorage,
+    {
+        let index_key = StorageKey::new(
+            evm.account_idxs.prefix(),
+            account,
+            evm.account_idxs.codec().key_codec(),
+        );
+        let index_proof = working_set.get_with_proof(index_key, version);
+        let index_proof_exists = index_proof.value.is_some();
+        let index_proof = borsh::to_vec(&index_proof.proof).expect("Serialization shouldn't fail");
+        let index_proof = Bytes::from(index_proof);
+
+        if index_proof_exists {
+            // we have to generate another proof for idx -> account
+            let index = evm
+                .account_idxs
+                .get(account, working_set)
+                .expect("Account index exists");
+            let index_bytes = Bytes::from_iter(index.to_le_bytes());
+
+            let account_key = StorageKey::new(
+                evm.accounts.prefix(),
+                &index,
+                evm.accounts.codec().key_codec(),
+            );
+
+            let account_proof = working_set.get_with_proof(account_key, version);
+            let account_exists = if account_proof.value.is_some() {
+                Bytes::from("y")
+            } else {
+                Bytes::from("n")
+            };
+            let account_proof =
+                borsh::to_vec(&account_proof.proof).expect("Serialization shouldn't fail");
+            let account_proof = Bytes::from(account_proof);
+            vec![index_proof, index_bytes, account_proof, account_exists]
+        } else {
+            let index_exists = Bytes::from("n");
+
+            vec![index_proof, index_exists]
+        }
+    }
+
+    fn generate_storage_proof<C>(
+        evm: &Evm<C>,
+        account: &Address,
+        key: JsonStorageKey,
+        version: u64,
+        working_set: &mut WorkingSet<C::Storage>,
+    ) -> EIP1186StorageProof
+    where
+        C: sov_modules_api::Context,
+        C::Storage: NativeStorage,
+    {
+        let key_b = key.as_b256().into();
+        let kaddr = Evm::<C>::get_storage_address(account, &key_b);
+        let storage_key = StorageKey::new(
+            evm.storage.prefix(),
+            &kaddr,
+            evm.storage.codec().key_codec(),
+        );
+        let value = evm.storage_get(account, &key_b, working_set);
+        let proof = working_set.get_with_proof(storage_key, version);
+        let value_exists = if proof.value.is_some() {
+            Bytes::from("y")
+        } else {
+            Bytes::from("n")
+        };
+        let value_proof = borsh::to_vec(&proof.proof).expect("Serialization shouldn't fail");
+        let value_proof = Bytes::from(value_proof);
+        EIP1186StorageProof {
+            key,
+            value: value.unwrap_or_default(),
+            proof: vec![value_proof, value_exists],
+        }
+    }
+
+    let account_proof = generate_account_proof(evm, &address, version, working_set);
+
+    let mut storage_proof = vec![];
+    for key in keys {
+        let proof = generate_storage_proof(evm, &address, key, version, working_set);
+        storage_proof.push(proof);
+    }
+
+    EIP1186AccountProofResponse {
+        address,
+        balance,
+        nonce,
+        code_hash,
+        storage_hash: root_hash.into(),
+        account_proof,
+        storage_proof,
     }
 }
